Fixed bugs related to missing images and usage of glob command on *nix
In *nix, the '\\' character is not equivalent to the '/' character like
in windows. So make sure when using the glob command, all paths are
specified with '/', or else the code will not work on macs.

Also, the image file used to initialize the prime_image object was
removed/relocated so now a different image is being used. The specific
image doesn't matter as long as it exists and has the same aspect ratio
as the prime images