Add: animations
Now visual.show_image_sequence accepts a wait_function parameter.
This function is called every time an image is shown with the path to the shown image as a parameter.
Added a default implementation, default_wait_function, which waits for a spacebar press to go to the next image,
or if the image name contains the word 'animate', waits a specified amount of time and goes to the next one.